ada beberapa perubahan code dan upload file baru

+++ admin/pesanan.php
@@ -50,7 +50,7 @@
                         <td>
                             <span>
                                 <a href="index.php?halaman=editPesanan&id=<?php echo $data['id_pesanan'] ?>" class="btn btn-warning">Ubah</a>
-                                <a href="hapusPes.php?id=<?php echo $data['id_pesanan'] ?>" class="btn btn-danger">Hapus</a>
+                                <a href="hapusPes.php?id=<?php echo $data['id_pesanan'] ?>" class="btn btn-danger" onclick="return confirm('Apakah Anda yakin ingin menghapus pesanan ini?');">Hapus</a>
                             </span>
                         </td>
